making contacts route protected

+++ controllers/contactController.js
@@ -2,11 +2,11 @@ import Contacts from "../models/contactModel.js"
 
 export const getContacts = async(req,res,next)=>{
     try {
-        const contacts = await Contacts.find();
+        const contacts = await Contacts.find({user_id : req.user.id});
         return res.status(200).json(contacts)
     } 
     catch (error) {
-        next(new Error(error.message))
+        next(error)
     }    
 }
 
@@ -38,7 +38,7 @@ export const createContact = async(req,res,next)=>{
         throw new Error("All Fields are mandatory")
     }
     const contact = await Contacts.create({
-        name, email, phone
+        name, email, phone , user_id:req.user.id
     })
     return res.status(201).json(contact)
     }
@@ -59,6 +59,10 @@ export const updateContact = async(req,res,next)=>{
             res.status(404)
             throw new Error("No contacts found")
         }
+        if(contact.user_id.toString() !== req.user.id){
+            res.status(403)
+            throw new Error("User is not authorised to perform this action")
+        }
         const updatedContact = await Contacts.findByIdAndUpdate(id, req.body,{new:true});
         return res.status(200).json(updatedContact)
     }
@@ -78,6 +82,10 @@ export const deleteContact = async(req,res,next)=>{
         if(!contact){
             res.status(404)
             throw new Error("No Contacts found");
+        }
+        if(contact.user_id.toString() !== req.user.id){
+            res.status(403)
+            throw new Error("User is not authorised to perform this action")
         }
         await Contacts.findByIdAndDelete(id);
         return res.status(200).json("Contact deleted")
